修改遥控器摇杆数据类型2

+++ Seekfree_CYT4BB_Opensource_Library/project/code/app/remote/remote_comm.h
@@ -5,9 +5,23 @@
 #include "../../common/types.h"
 
 typedef struct {
-    float joystick[4];
+    int16 joystick[4];
+    // 0-左边 左右
+    // 1-左边 上下
+    // 2-右边 左右
+    // 3-右边 上下
+
     uint8 key[4];
+    // 0-摇杆左边
+    // 1-摇杆右边
+    // 2-侧向按键左边
+    // 3-侧向按键右边
+
     uint8 switch_key[4];
+    // 0-左边_1
+    // 1-左边_2
+    // 2-右边_1
+    // 3-右边_2
     bool connected;
     bool frame_updated;
 } Remote_State_t;
